pyproject.toml
[tool.poetry]
name = "pythondaq"
version = "0.1.0"
description = ""
authors = ["RJVB09 <rjvb09@ziggo.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
matplotlib = "^3.9.2"
pyvisa-py = "^0.7.2"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
